app/requirements.txt
fastapi
uvicorn
python-multipart
jinja2
itsdangerous

langchain
langchain-community
langchain-core
langchain-openai
langchain-chroma
langchain-text-splitters
chromadb

pymupdf
python-docx
python-dotenv
markdown